requirements.txt
psycopg2-binary>=2.9.9
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
